+++ changelog.md
@@ -7,5 +7,6 @@
 - Added the ability to create a new workout file
 - Added the ability to select a stored workout and view it
 - Added error message when starting an empty workout
+- Added option to add new steps to a workout
 - redundant `view workout` button removed
 - Small UI updates